Revert "Merge branch 'day_night_system' of https://github.com/quocmanh94/pokeemerald into PKMN-Emerald-X2"
This reverts commit 515beea83c0b7f40ae65b495e07da844ddb0e459, reversing
changes made to 8d77fad6efc3cc8b9724be8379cc658568c8f0ce.

+++ src/battle_tower.c
@@ -1257,11 +1257,11 @@ void SetEReaderTrainerGfxId(void)
     SetBattleFacilityTrainerGfxId(TRAINER_EREADER, 0);
 }
 
-u16 GetBattleFacilityTrainerGfxId(u16 trainerId)
+u8 GetBattleFacilityTrainerGfxId(u16 trainerId)
 {
     u32 i;
     u8 facilityClass;
-    u16 trainerObjectGfxId;
+    u8 trainerObjectGfxId;
 
     SetFacilityPtrsGetLevel();
     if (trainerId == TRAINER_EREADER)
@@ -3584,7 +3584,7 @@ static void FillTentTrainerParty_(u16 trainerId, u8 firstMonId, u8 monCount)
     }
 }
 
-u16 FacilityClassToGraphicsId(u8 facilityClass)
+u8 FacilityClassToGraphicsId(u8 facilityClass)
 {
     u8 trainerObjectGfxId;
     u8 i;
